requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docopt==0.6.2
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Sessionstore==0.4.5
Flask-WTF==0.14.2
fuzzywuzzy==0.17.0
get==2019.4.13
graphqlclient==0.2.4
idna==2.8
inexactsearch==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
pipreqs==0.4.9
post==2019.4.13
public==2019.4.13
pyspellchecker==0.4.0
PyStemmer==1.3.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
silpa-common==0.3
six==1.12.0
sklearn==0.0
soundex==1.1.3
urllib3==1.24.2
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
yarg==0.1.9
